edit user permission

+++ server/routes/expense.js
@@ -62,26 +62,47 @@ expenseRouter.route('/:expenseId')
   .post(authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end('POST operation not supported on /expenses/' + req.params.expenseId);
+
   })
   .put(authenticate.verifyUser, (req, res, next) => {
-    Expenses.findByIdAndUpdate(req.params.expenseId, {
-      $set: req.body
-    }, { new: true })
+    // check if the user is the owner of the expense
+    Expenses.findById(req.params.expenseId)
       .then((expense) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(expense);
-      }, (err) => next(err))
-      .catch((err) => next(err));
+        if (expense.user.equals(req.user._id)) {
+          Expenses.findByIdAndUpdate(req.params.expenseId, {
+            $set: req.body
+          }, { new: true })
+            .then((expense) => {
+              Expenses.findById(expense._id)
+                .populate('category')
+                .then((expense) => {
+                  res.statusCode = 200;
+                  res.setHeader('Content-Type', 'application/json');
+                  res.json(expense);
+                })
+            }, (err) => next(err))
+            .catch((err) => next(err));
+        }
+      })
   })
   .delete(authenticate.verifyUser, (req, res, next) => {
-    Expenses.findByIdAndRemove(req.params.expenseId)
-      .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-      }, (err) => next(err))
-      .catch((err) => next(err));
+    // check if the user is the owner of the expense
+    Expenses.findById(req.params.expenseId)
+      .then((expense) => {
+        if (expense.user.equals(req.user._id)) {
+          Expenses.findByIdAndRemove(req.params.expenseId)
+            .then((resp) => {
+              res.statusCode = 200;
+              res.setHeader('Content-Type', 'application/json');
+              res.json(resp);
+            }, (err) => next(err))
+            .catch((err) => next(err));
+        } else {
+          var err = new Error('You are not authorized to delete this expense!');
+          err.status = 403;
+          return next(err);
+        }
+      })
   });
 
 module.exports = expenseRouter;
